backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.28.0
anthropic>=0.52.0
python-dotenv>=1.0.0
slowapi>=0.1.9
PyJWT[crypto]>=2.8.0
